crypto/Cube_Pyramid/requirements.txt
pycryptodome==3.20.0
sympy==1.13.1